Add a Rails compatible TaggedLogger
I've split this out to its own Gem b/c it has a dependency on
`ActiveSupport`'s `Logger` and `TaggedLogger` constants. This is a bit
cumbersome as we need to maintain two different Gems, and probably
deploy them in lock step. Alternatively we could probably accomplish
this by way of documentation (telling folks to add `activesupport`) to
their Gemfile/.gemspec if they want to leverage the tagged logging
stuff. I'm kinda 50/50 on it, so… 🤷